Post model's methods checked. Model's index and creation pages added

+++ database/migrations/2019_09_10_063615_create_characteristic_translations_table.php
@@ -17,7 +17,7 @@ class CreateCharacteristicTranslationsTable extends Migration
             $table->bigIncrements('id');
 
             // 'id' field on characteristics table
-            $table->integer('characteristic_id')->unsigned();
+            $table->bigInteger('characteristic_id')->unsigned();
             $table->string('locale')->index();
 
             $table->string('key')->nullable();
